Add runtime instrumentation (§32); finalise audit documentation
OperationTimings records count/total/max and bucketed p50/p95/p99 per named
operation, so the performance claims in the report can be checked on the machine
actually running the application rather than trusted from one benchmark host.

- Bounded memory: ~40 logarithmic buckets, not retained samples. Never
  under-reports; never over-reports by more than 2x.
- Under 2 us per call, asserted as a test budget, so measuring does not distort
  what is measured. Exact under eight concurrent writers. Disableable via
  -Daegis.instrumentation=false.
- Wired into DashboardStats.snapshot/rebuild, LuceneIndex.put/search and
  SearchFacets.counts; DB write, relationship, extraction and OCR constants are
  declared for the paths still unprofiled.

Docs: FINAL_STATUS gains a directive checklist, headline results and an explicit
gaps section; DATABASE_PERFORMANCE_REPORT gains the instrumentation section and a
test-status section; coverage inventory grows to 99 rows.

Headless suite 160/160 pass; InterfaceFunctionMatrixTest 7/7.

+++ app/src/main/java/com/aegis/fdx/ui/screens/DashboardScreen.java
@@ -125,7 +125,7 @@ public final class DashboardScreen implements Screen {
         }
     }
 
-    private static Node statusRow(String label, int value, String color) {
+    private static Node statusRow(String label, long value, String color) {
         Label l = new Label(label);
         l.setStyle("-fx-font-size: 12px; -fx-text-fill: " + Fas.TEXT_DARK + ";");
         Label v = new Label(String.format("%,d", value));
